Move methods from app.rb to module file
